solverServer/requirements.txt
astrometry==4.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
paho-mqtt==2.1.0
requests==2.32.3
urllib3==2.2.3
